Feat: 데이터 파이프라인 가동 — 크롤러/프로세서 진입점 및 시드 데이터 - backend/crawler/main.py: APScheduler 기반 크롤러 서비스 진입점 - backend/processor/main.py: 미처리 기사 polling 프로세서 진입점 - migrations/014: source_config 시드 + feed_source 소급 backfill + admin_settings 시드 - update_feed_health 쿼리 float8 캐스트 수정 - 크롤러/프로세서 모두 Redis init_redis() 호출 추가 Closes: #58

+++ backend/db/queries/feed_sources.py
@@ -297,8 +297,8 @@ async def update_feed_health(
                         last_success_at = now(),
                         consecutive_failures = 0,
                         avg_latency_ms = CASE
-                            WHEN avg_latency_ms IS NULL THEN $2
-                            ELSE (avg_latency_ms * 0.7 + $2 * 0.3)
+                            WHEN avg_latency_ms IS NULL THEN $2::float8
+                            ELSE (avg_latency_ms * 0.7 + $2::float8 * 0.3)
                         END,
                         total_crawl_count = total_crawl_count + 1,
                         updated_at = now()
@@ -316,8 +316,8 @@ async def update_feed_health(
                         last_error_at = now(),
                         consecutive_failures = consecutive_failures + 1,
                         avg_latency_ms = CASE
-                            WHEN avg_latency_ms IS NULL THEN $3
-                            ELSE (avg_latency_ms * 0.7 + $3 * 0.3)
+                            WHEN avg_latency_ms IS NULL THEN $3::float8
+                            ELSE (avg_latency_ms * 0.7 + $3::float8 * 0.3)
                         END,
                         total_crawl_count = total_crawl_count + 1,
                         total_error_count = total_error_count + 1,
